Move delegating primitive into internal package.
