+++ core/src/main/java/io/atomix/core/map/impl/DelegatingAsyncAtomicMap.java
@@ -24,7 +24,7 @@ import io.atomix.core.map.AtomicMap;
 import io.atomix.core.map.AtomicMapEventListener;
 import io.atomix.core.transaction.TransactionId;
 import io.atomix.core.transaction.TransactionLog;
-import io.atomix.primitive.DelegatingAsyncPrimitive;
+import io.atomix.primitive.impl.DelegatingAsyncPrimitive;
 import io.atomix.primitive.PrimitiveState;
 import io.atomix.utils.time.Versioned;
