+++ core/src/main/java/io/atomix/core/tree/impl/DelegatingAsyncAtomicDocumentTree.java
@@ -20,7 +20,7 @@ import io.atomix.core.tree.AsyncAtomicDocumentTree;
 import io.atomix.core.tree.DocumentPath;
 import io.atomix.core.tree.AtomicDocumentTree;
 import io.atomix.core.tree.DocumentTreeListener;
-import io.atomix.primitive.DelegatingAsyncPrimitive;
+import io.atomix.primitive.impl.DelegatingAsyncPrimitive;
 import io.atomix.utils.time.Versioned;
 
 import java.time.Duration;
